refactor topic messages  with firebase

+++ presentation/viewmodel/src/main/java/com/chocolate/viewmodel/saveLater/SaveLaterMessageUiState.kt
@@ -22,7 +22,7 @@ data class MessageItemUiState(
 
 fun Message.toMessageUiState(): MessageItemUiState {
     return MessageItemUiState(
-        id = id,
+        id = id.toInt(),
         username = senderFullName,
        imageUrl = senderAvatarUrl,
         messageContent = messageContent,
